More -ansi -pedantic cleanup.
git-svn-id: file:///home/shlomif/Backup/svn-dumps/google-code/svnsync-repos/fc-solve/trunk@541 e7e8a897-7ba4-4ee7-b36f-f4c66519b19a

+++ fc-solve/source/alloc.h
@@ -36,7 +36,7 @@ extern char *
 #define fcs_compact_alloc_into_var(result,allocator_orig,what_t) \
 { \
    register fcs_compact_allocator_t * allocator = (allocator_orig); \
-   if (allocator->max_ptr - allocator->ptr < sizeof(what_t))  \
+   if (allocator->max_ptr - allocator->ptr < ((int)sizeof(what_t)))  \
     {      \
         freecell_solver_compact_allocator_extend(allocator);      \
     }         \
